refactor(maker): 拆分统计告警与大型测试模块
收益: 将 MakerStats、AlertMonitor 与两个千行级内联测试模块拆到聚焦文件，lib.rs 从 3159 行降至约 1007 行。

风险: maker 安全与交易语义必须保持不变；脚本逐字节确认实现和测试仅移动/去缩进，根级 API 继续 re-export，190 个 maker 测试通过，独立对抗审查未发现高严重度问题。
